xlstream-eval: rename regression test files for clarity
